requirements.txt
mysql-connector-python==8.0.33
protobuf==3.20.3
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
